+++ events/views.py
@@ -1,13 +1,15 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
 from .serializers.populated import PopulatedEventSerializer
 from .serializers.common import EventSerializer
 from .models import Event
 
 class EventListView(APIView):
+    permission_classes = (IsAuthenticatedOrReadOnly,)
 
     def get(self, _request):
         events = Event.objects.all()
@@ -15,6 +17,7 @@ class EventListView(APIView):
         return Response(serialized_events.data, status=status.HTTP_200_OK)
 
     def post(self, request):
+        request.data["owner"] = request.user.id
         event_to_add = EventSerializer(data=request.data)
         if event_to_add.is_valid():
             event_to_add.save()
@@ -23,6 +26,7 @@ class EventListView(APIView):
 
 
 class EventDetailView(APIView):
+    permission_classes = (IsAuthenticatedOrReadOnly,)
 
     def get_event(self, pk):
         try:
@@ -35,14 +39,19 @@ class EventDetailView(APIView):
         serialized_event = EventSerializer(event)
         return Response(serialized_event.data, status=status.HTTP_200_OK)
 
-    def delete(self, _request, pk):
+    def delete(self, request, pk):
         event_to_delete = self.get_event(pk=pk)
+        if event_to_delete.owner != request.user:
+            raise PermissionDenied()
         event_to_delete.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     def put(self, request, pk):
         event_to_edit = self.get_event(pk=pk)
         updated_event = EventSerializer(event_to_edit, data=request.data)
+        request.data["owner"] = request.user.id
+        if event_to_edit.owner != request.user:
+            raise PermissionDenied()
         if updated_event.is_valid():
             updated_event.save()
             return Response(updated_event.data, status=status.HTTP_202_ACCEPTED)
